% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOTPLY.R
\name{select_top_weights}
\alias{select_top_weights}
\title{select_top_weights}
\usage{
select_top_weights(transition_matrix, top_link = NULL)
}
\arguments{
\item{transition_matrix}{Output from \emph{get_transition_matrix}.}

\item{top_link}{Integer.Maximum number of links to select between clusters at time t and clusters at time t+1. Links are sorted according to the weigth and then only the \emph{top_link} are kept. If \emph{NULL} (default), all the links are kept.}
}
\value{
A matrix
}
\description{
select_top_weights
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
