% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transition_matrix.R
\name{convert_names}
\alias{convert_names}
\title{convert_names}
\usage{
convert_names(new_row, new_col, transition_matrix)
}
\arguments{
\item{new_row}{Vector with the new row names to assign to \emph{transition_matrix}}

\item{new_col}{Vector with the new column names to assign to \emph{transition_matrix}}

\item{transition_matrix}{Output from \emph{get_transition_matrix}.}
}
\value{
A matrix with row names equal to \emph{new_row} and column names eqaul to
\emph{new_col}.
}
\description{
convert_names
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
